.history/requirements_20250524114116.txt
streamlit
pandas
mlxtend
mysql-connector-python
